refactor!: move internal utils functions to a dedicated directory
Such functions are now easier to identify, so this should help for the maintenance and should prevent from exporting
them by mistake.

BREAKING CHANGES: `Utils.copyTextToClipboard` is no longer available. It was intended to be internal and had been made public by mistake.
